Correct a typo

+++ 0x08-recursion/2-strlen_recursion.c
@@ -9,5 +9,5 @@
 int _strlen_recursion(char *str)
 {
 	if (*str != '\0')
-		return (1 + _strlen_recursion(str + 1));
+		return (int) (1 + _strlen_recursion(str + 1));
 }
